+++ src/bika/lims/browser/analysisrequest/add2.py
@@ -1907,6 +1907,15 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
                     msg = _("Contact does not belong to the selected client")
                     fielderrors["Contact"] = msg
 
+            # Auto-add CCContact when hidden on Sample Add form
+            field = self.get_field("CCContact")
+            hidden_fields = self.get_fields_with_visibility(
+                    "hidden", mode="add")
+            if field in hidden_fields and contact_obj:
+                cc_contacts = contact_obj.getCCContact()
+                if cc_contacts:
+                    record["CCContact"] = [cc.UID() for cc in cc_contacts]
+
             # Check if the number of samples per record is permitted
             num_samples = self.get_num_samples(record)
             if num_samples > max_samples_record: